requirement_deepsort.txt
openvino-dev==2023.0.0
deep_sort_realtime
numpy
scipy
dataclasses
opencv-python
Pillow
torch
torchvision
typing-extensions